# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZK-VOTE Relayer API",
    "version": "1.0.0",
    "description": "Backend relayer for anonymous voting on Stellar/Soroban. Generated from route definitions and Zod validation schemas — see backend/API.md for prose docs and GET /api-docs for interactive documentation."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "RelayerAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Relayer-Auth",
        "description": "Shared relayer auth token (also accepted as `Authorization: Bearer <token>`). There is currently one shared token per deployment, not per-user credentials."
      }
    },
...
  },
  "paths": {
...
    "/db/stats": {
      "get": {
        "tags": [
          "Health"
        ],
        "summary": "Database diagnostics (full detail requires auth)",
        "description": "No rate limit.",
        "security": [
          {
            "RelayerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {},
                "example": {
                  "queries": {},
                  "tables": [],
                  "cache": {}
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}